// microInventories/src/types/inventories.type.ts
export interface Inventory {
	inventory_id: number
	user_id: number;
	product_id: number,
	product_name:string,
	product_image:string,
	rarity_id:number,
	rarity_name:string,
	created_at: string;
	updated_at: string;
}

export interface InventoryRaw {
	inventory_id:number;
	user_id: number;
	product_id: number;
	created_at: string;
	updated_at: string;
}